requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.6
distro==1.9.0
Django==5.2.5
django-encrypted-model-fields==0.6.5
django-picklefield==3.3
django-q2==1.8.0
gevent==25.5.1
greenlet==3.2.4
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8.2
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.2
openai==1.99.9
pandas==2.3.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
slack_sdk==3.36.0
slackify_markdown==0.2.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.event==5.1.1
zope.interface==7.2
